Add files via upload
In Python console, type : pip install Pillow

This updated script includes the get_date_taken function, which uses the Pillow library to extract the date taken from image metadata. The script then organizes the photos based on this date taken value. If the date taken information is not available for a file, it will be skipped and a message will be printed.